+++ lib/models/page.dart
@@ -5,7 +5,7 @@ import "dart:html";
 import "package:polymer/polymer.dart";
 import 'package:route_hierarchical/client.dart';
 
-class Page extends JsProxy {
+class AppPage extends JsProxy {
 
   /// path of the page
   @reflectable
@@ -18,7 +18,7 @@ class Page extends JsProxy {
   /// element of the page
   /// Can be an [HtmlElement] or the element name as a [String]
   @reflectable
-  final dynamic element;
+  dynamic element;
 
   /// definne if page is default, home
   @reflectable
@@ -32,12 +32,16 @@ class Page extends JsProxy {
   @reflectable
   final hideLeftNav;
 
+  /// define polymer material icon for the page
+  @reflectable
+  dynamic icon;
+
   /// link to the sub page
   @reflectable
-  Page child;
+  AppPage child;
 
   String toString() =>
-      "{ name: $name, path: $path, element: $element, isDefault: $isDefault, menu: $menu, hideLeftNav: $hideLeftNav}";
+      "{ name: $name, path: $path, element: $element, isDefault: $isDefault, menu: $menu, hideLeftNav: $hideLeftNav, icon: $icon}";
 
   /// When enter in a new route.
   /// This function call the enterRoute function of the current element if it's possible.
@@ -50,6 +54,19 @@ class Page extends JsProxy {
     }
   }
 
-  Page(this.name, this.path, this.element,
-      {this.child: null, this.isDefault: false, this.menu: true, this.hideLeftNav: false});
+  AppPage(this.name, this.path, _element, {this.isDefault: false, this.menu: true, this.hideLeftNav: false, this.icon, this.child: null}) {
+    if (icon is String || icon is HtmlElement) {
+      this.icon = icon;
+    } else {
+      this.icon = null;
+    }
+    if (_element is String) {
+      element = document.createElement(_element);
+    } else if (_element is HtmlElement) {
+      element = _element;
+    } else {
+      throw "Page : element must be String or HtmlElement.";
+    }
+    this.child = child;
+  }
 }