+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.5
+
+### Fix loading for dart 1.13
+### Breaking change
+* Replace Page by AppPage
+### Update demo
+
+
 ## 0.1.4
 
 ### Fix NavHeader when it is custome